lending pool function added but test failing

+++ src/mocks/mockErc20.sol
@@ -6,11 +6,11 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 contract MockERC20 is ERC20 {
     constructor(string memory name, string memory symbol) ERC20(name, symbol) {}
 
-    function mint(address account, uint amount) external {
+    function mint(address account, uint256 amount) external {
         _mint(account, amount);
     }
 
-    function burn(address account, uint amount) external {
+    function burn(address account, uint256 amount) external {
         _burn(account, amount);
     }
 }
